Addressed Issue #3 -- Creating Many Rows in Inventory with the Same user_id/badge_id combination.
How to update a row:

I was able to increment the number of badges a user has by doing the
following in the rails console.

Select the inventory where user_id is 1 and badge_id is 1.
i = Inventory.lock.where(:user_id => '1', :badge_id => '1')

Lock down i and send it to k. I read it on one of the ruby guide pages
and it worked.
k = i.lock.first

Increment the value in :amount by 1. You could put it in a loop just in
case the user wants to increment by any number.
k.increment!(:amount)

I left what I put above as a comment in the inventory.rb that I
committed just now.