Increment button functions when pressed
This is begging for AJAX calls but I just could not get it to work out.
Right now the onClick will submit the form as well as quickly refresh
the page. If you delay the clicks by a second, the page refreshes so
quickly it is not noticeable.

To do: Should we limit the max amount of XP to the sum of all game
levels xp? Right now user Ron has 48 XP with only 15 xp on level 2.